+++ dart/geobase/CHANGELOG.md
@@ -28,6 +28,7 @@ binary data formats, encodings and content interfaces also redesigned.
 - [Equals, hashCode and toString for geometries and features on geobase #143](https://github.com/navibyte/geospatial/issues/143)
 - [Encode / write methods directly from geometry and feature classes on geobase #144](https://github.com/navibyte/geospatial/issues/144)
 - [Decode GeoJSON and build data structures using new data structures by geobase #140](https://github.com/navibyte/geospatial/issues/140)
+- [Handle explicit and implicit bounds in geometries and features on geobase](https://github.com/navibyte/geospatial/issues/141)
 
 🛠 Refactoring:
 - [Optimize reading from iterable of coordinate values #131](https://github.com/navibyte/geospatial/issues/131)